app/requirements.txt
pymongo
pandas-datareader
pandas
numpy
tensorflow
sklearn
python-dotenv
flask-restful